Use Gatsby dev to proxy dev-server
