+++ develop.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+trap 'kill 0' EXIT
+
+npm run server &
+cd static-site && npm run develop
+
+wait
