+++ package.json
@@ -19,7 +19,8 @@
     "start": "npm run server",
     "smoke-test": "NODE_ENV=test ENV=dev jest ./test/smoke-test/auspice_client_requests.test.js",
     "smoke-test:ci": "start-server-and-test server http://localhost:5000 smoke-test",
-    "test:ci": "npm run smoke-test:ci"
+    "test:ci": "npm run smoke-test:ci",
+    "dev": "./develop.sh"
   },
   "dependencies": {
     "argparse": "^1.0.10",
