+++ server.js
@@ -8,6 +8,7 @@ const favicon = require('serve-favicon');
 const compression = require('compression');
 const argparse = require('argparse');
 const utils = require("./src/utils");
+const auspicePaths = require('./auspicePaths');
 const cors = require('cors');
 
 const production = process.env.NODE_ENV === "production";
@@ -34,7 +35,6 @@ global.verbose = args.verbose;
 // in them can use utils.verbose() at load time.
 const auspiceServerHandlers = require("./src/index.js");
 const authn = require("./authn");
-const sources = require("./src/sources");
 const redirects = require("./redirects");
 
 /* Path helpers for static assets, to make routes more readable.
@@ -100,52 +100,6 @@ app.route("/charon/getSourceInfo")
 
 app.route("/charon/*")
   .all((req, res) => res.status(404).end());
-
-
-/* Dataset and narrative paths hit Auspice's entrypoint.
- */
-const coreBuilds = [
-  "/dengue",
-  "/ebola",
-  "/enterovirus",
-  "/flu",
-  "/lassa",
-  "/measles",
-  "/mers",
-  "/mumps",
-  "/ncov",
-  "/tb",
-  "/WNV",
-  "/yellow-fever",
-  "/zika",
-];
-
-const groups = [...sources.keys()].filter((k) => !["core", "staging", "community"].includes(k));
-
-const auspicePaths = [
-  "/status",
-  "/status/*",
-  ...coreBuilds,
-  ...coreBuilds.map((url) => url + ":*"),
-  ...coreBuilds.map(url => url + "/*"),
-  "/narratives",
-  "/narratives/*",
-  "/staging",
-  "/staging/*",
-  ...groups.map((group) => `/groups/${group}`),
-  ...groups.map((group) => `/groups/${group}/*`),
-
-  /* Auspice gets specific /community paths so it can show an index of datasets
-   * and narratives, but Gatsby gets top-level /community.
-   */
-  "/community/:user/:repo",
-  "/community/:user/:repo/*",
-
-  /* auspice gets /fetch/X URLs which result in loading of dataset accessible at URL X
-   * note that gatsby will redirect /fetch to the docs page explaining this behavior
-   */
-  "/fetch/*"
-];
 
 app.route(auspicePaths).get((req, res) => {
   utils.verbose(`Sending Auspice entrypoint for ${req.originalUrl}`);
