+++ static-site/gatsby-config.js
@@ -1,4 +1,6 @@
+const createProxyMiddleware = require("http-proxy-middleware");
 const config = require("./data/SiteConfig");
+const auspicePaths = require("../auspicePaths");
 
 // const pathPrefix = config.pathPrefix === "/" ? "" : config.pathPrefix;
 
@@ -69,5 +71,15 @@ module.exports = {
     },
     "gatsby-plugin-sharp",
     // "gatsby-plugin-catch-links" // See https://github.com/nextstrain/nextstrain.org/issues/34
-  ]
+  ],
+  developMiddleware: app => {
+    ['/charon', '/dist', ...auspicePaths].forEach(path => {
+      app.use(
+        path,
+        createProxyMiddleware({
+          target: "http://localhost:5000",
+        })
+      );
+    })
+  },
 };
